Add addPrerequisite method to ProgramRepository
